% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_trajectory.R
\name{choose_regression_model}
\alias{choose_regression_model}
\title{Find a regression method (linear, lasso, or ridge)}
\usage{
choose_regression_model(model_name)
}
\arguments{
\item{model_name}{Name of the model}
}
\value{
A model specification
}
\description{
Find a regression method (linear, lasso, or ridge)
}
