% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_trajectory.R
\name{differentiate_ts}
\alias{differentiate_ts}
\title{Get the deriative of a time series}
\usage{
differentiate_ts(ts)
}
\arguments{
\item{ts}{Time series data}
}
\value{
A tibble with original speecies abundance and its first derivative
}
\description{
Get the deriative of a time series
}
