% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_trajectory.R
\name{evaluate_fit}
\alias{evaluate_fit}
\title{Evaluate how close the simulated time series with fitted parameters is to true time series}
\usage{
evaluate_fit(ts, ts_simu)
}
\value{
value of NRMSE
}
\description{
Evaluate how close the simulated time series with fitted parameters is to true time series
}
