% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_trajectory.R
\name{fit_interaction_parameters}
\alias{fit_interaction_parameters}
\title{Fit interaction matrix with given topology and intrinsic growth rates from time series}
\usage{
fit_interaction_parameters(
  ts_species,
  reg_model = choose_regression_model("linear"),
  topology_all
)
}
\value{
A tibble with fitted parameters
}
\description{
Fit interaction matrix with given topology and intrinsic growth rates from time series
}
