% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_trajectory.R
\name{fit_parameters}
\alias{fit_parameters}
\title{Fit interaction matrix with given topology and intrinsic growth rates from time series}
\usage{
fit_parameters(ts, topology)
}
\arguments{
\item{Sigma}{Interaction matrix}

\item{r}{intrinsic growth rates}
}
\value{
A tibble with simulated time series of species abundances
}
\description{
Fit interaction matrix with given topology and intrinsic growth rates from time series
}
