% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_trajectory.R
\name{generate_top_model_candidates}
\alias{generate_top_model_candidates}
\title{Get the combinatories of models from highest R^2 from each species}
\usage{
generate_top_model_candidates(fitted_models, num_each_species = 3)
}
\arguments{
\item{fitted_models}{Specific topology}

\item{num_each_species}{How many topologies to choose for each species}
}
\value{
A tibble with top model candidates
}
\description{
Get the combinatories of models from highest R^2 from each species
}
