% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equations.R
\name{get_classic_dynamics}
\alias{get_classic_dynamics}
\title{Get the parameters for some classic population dynamics}
\usage{
get_classic_dynamics(
  equation_name,
  state_initial,
  time_range,
  species_num,
  topology_ground
)
}
\arguments{
\item{equation_name}{Name of the classic population dynamics}
}
\description{
Get the parameters for some classic population dynamics
}
