% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_trajectory.R
\name{get_topology_label}
\alias{get_topology_label}
\title{Get the label of the specific topology}
\usage{
get_topology_label(topology, topology_all)
}
\arguments{
\item{topology}{Specific topology}

\item{topology_all}{All possible topology that are used to fit}
}
\value{
A ggraph project
}
\description{
Get the label of the specific topology
}
