% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_trajectory.R
\name{plot_interaction_topology}
\alias{plot_interaction_topology}
\title{EPlot the interaction network}
\usage{
plot_interaction_topology(topology)
}
\value{
A ggraph project
}
\description{
EPlot the interaction network
}
