% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_ODEs.R
\name{plot_time_series}
\alias{plot_time_series}
\title{Plot time series of species abundance}
\usage{
plot_time_series(dataset)
}
\value{
A ggplot2 object
}
\description{
Plot time series of species abundance
}
