% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_trajectory.R
\name{plot_true_vs_simu}
\alias{plot_true_vs_simu}
\title{Plot simulated time series with fitted parameters vs true time series}
\usage{
plot_true_vs_simu(ts, ts_simu)
}
\value{
A ggplot2 object
}
\description{
Plot simulated time series with fitted parameters vs true time series
}
