% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_trajectory.R
\name{preprocess_ts}
\alias{preprocess_ts}
\title{Transform a time series data into a tibble that is ready to fit the regression}
\usage{
preprocess_ts(ts)
}
\arguments{
\item{ts}{Time series data}
}
\value{
A tibble with log change of species abundance with original abundance
}
\description{
Transform a time series data into a tibble that is ready to fit the regression
}
