% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_trajectory.R
\name{simualte_fitted_dynamics}
\alias{simualte_fitted_dynamics}
\title{Simulate time series from fitted parameters}
\usage{
simualte_fitted_dynamics(topology_fitted)
}
\arguments{
\item{topology_fitted}{The fitted parameter of the dynamics}
}
\value{
A tibble with simulated time series of species abundances
}
\description{
Simulate time series from fitted parameters
}
