% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InferInteractions-package.R
\docType{package}
\name{topodetective-package}
\alias{topodetective}
\alias{topodetective-package}
\title{topodetective: Infer topology of species interactions from time series of species abundance}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Chuliang Song \email{clsong.ecology@gmail.com} (\href{https://orcid.org/0000-0001-7490-8626}{ORCID})

}
\keyword{internal}
