% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_ODEs.R
\name{generate_time_series_LV}
\alias{generate_time_series_LV}
\title{Simulated time series of species abundances}
\usage{
generate_time_series_LV(topology, state_initial, time_range)
}
\arguments{
\item{state_initial}{Initial species abundances}

\item{time_range}{Time range to run the simulation}

\item{Sigma}{Interaction matrix}

\item{r}{intrinsic growth rates}
}
\value{
A tibble with simulated time series of species abundances
}
\description{
Simulated time series of species abundances
}
