% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equations.R
\name{get_LV_dynamics}
\alias{get_LV_dynamics}
\title{Get the parameters for the three phases of LV dynamics following Bunin (2017) PRE}
\usage{
get_LV_dynamics(species_num, conne, mu, sigma)
}
\arguments{
\item{species_num}{Number of species}
}
\description{
Get the parameters for the three phases of LV dynamics following Bunin (2017) PRE
}
