% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_interaction_topology}
\alias{plot_interaction_topology}
\title{Plot the interaction network and intrinsic growth rates}
\usage{
plot_interaction_topology(topology, title)
}
\arguments{
\item{topology}{A tibble with species intrinsic growth rates and interaction strength}

\item{title}{title of the ggplot}
}
\value{
A ggraph project
}
\description{
Plot the interaction network and intrinsic growth rates
}
