% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_time_series}
\alias{plot_time_series}
\title{Plot time series of species abundance}
\usage{
plot_time_series(ts)
}
\arguments{
\item{ts}{A tibble containing time series}
}
\value{
A ggplot2 object
}
\description{
Plot time series of species abundance
}
